+++ node/src/types/kernel/api/storage/setLocalStorage/payload.d.ts
@@ -21,7 +21,7 @@
  */
 export interface IPayload {
     /**
-     * block ID
+     * App ID
      */
     app?: string;
     val:  { [key: string]: any };